try 2

+++ sql/item_func.h
@@ -778,7 +778,7 @@ class Item_func_hash :public Item_int_func
 public:
   Item_func_hash(THD *thd, Item *a): Item_int_func(thd, a)
   {}
-  int val_int();
+  longlong val_int();
   const char *func_name() const { return "HASH"; }
 };
